Rename package 'vika_kovalenko' to 'kovalenko.vika' and moved all classes; added constant variables to CaesarsCipher class
