+++ src/ua/com/javarush/kovalenko/vika/cryptoanalyser/basic/Main.java
@@ -1,0 +1,9 @@
+package ua.com.javarush.kovalenko.vika.cryptoanalyser.basic;
+
+import ua.com.javarush.kovalenko.vika.cryptoanalyser.model.Dialog;
+
+public class Main {
+    public static void main(String[] args) {
+        Dialog.dialogWithUser();
+    }
+}
