+++ src/ua/com/javarush/kovalenko/vika/cryptoanalyser/model/Alphabet.java
@@ -1,4 +1,4 @@
-package ua.com.javarush.vika_kovalenko.cryptoanalyser.model;
+package ua.com.javarush.kovalenko.vika.cryptoanalyser.model;
 
 public class Alphabet {
