+++ src/ua/com/javarush/kovalenko/vika/cryptoanalyser/model/Dialog.java
@@ -1,7 +1,7 @@
-package ua.com.javarush.vika_kovalenko.cryptoanalyser.model;
+package ua.com.javarush.kovalenko.vika.cryptoanalyser.model;
 
-import ua.com.javarush.vika_kovalenko.cryptoanalyser.files.CaesarsCipher;
-import ua.com.javarush.vika_kovalenko.cryptoanalyser.files.FileChecking;
+import ua.com.javarush.kovalenko.vika.cryptoanalyser.files.CaesarsCipher;
+import ua.com.javarush.kovalenko.vika.cryptoanalyser.files.FileChecking;
 
 import java.nio.file.InvalidPathException;
 import java.nio.file.Path;
