--- src/ua/com/javarush/vika_kovalenko/cryptoanalyser/basic/Main.java
@@ -1,9 +1,0 @@
-package ua.com.javarush.vika_kovalenko.cryptoanalyser.basic;
-
-import ua.com.javarush.vika_kovalenko.cryptoanalyser.model.Dialog;
-
-public class Main {
-    public static void main(String[] args) {
-        Dialog.dialogWithUser();
-    }
-}
